Splitting display to new file
