+++ src/common.h
@@ -50,12 +50,8 @@ void setConfLocations();
 void printVersion(char* programName);
 char * dirFromPath(const char* myStr);
 char * objectFromPath(const char* myStr);
-void wPrintLine(int line, int col, wchar_t *textString);
-void printLine(int line, int col, char *textString);
-void topLineMessage(const char *message);
 int check_dir(char *pwd);
 int check_file(char *file);
-void clear_workspace();
 int check_exec(const char *object);
 int check_last_char(const char *str, const char *chk);
 int check_first_char(const char *str, const char *chk);
