+++ src/display.h
@@ -1,0 +1,4 @@
+void wPrintLine(int line, int col, wchar_t *textString);
+void printLine(int line, int col, char *textString);
+void topLineMessage(const char *message);
+void clear_workspace();
